Refactor landlords and properties migrations
- Added 'name' field to landlords table and made 'phone' field optional.
- Changed 'management_fee' column type from DOUBLE PRECISION to DECIMAL(18, 2) in properties table.
- Removed obsolete migration for changing management_fee type.

+++ migrations/20250514001603_properties.sql
@@ -7,7 +7,7 @@ CREATE TABLE properties (
     suburb TEXT NOT NULL,
     postcode TEXT NOT NULL,
     state TEXT NOT NULL,
-    management_fee DOUBLE PRECISION NOT NULL,
+    management_fee DECIMAL(18, 2) NOT NULL,
     is_archived TIMESTAMP WITH TIME ZONE,
     landlord_id UUID NOT NULL REFERENCES landlords(id)
 );
